move ColliderFromMeshDef
