+++ src/wind/mast.ts
@@ -8,7 +8,7 @@ import { V, mat3, quat, tV, vec3 } from "../matrix/sprig-matrix.js";
 import { MastMesh } from "../meshes/mesh-list.js";
 import { LinearVelocityDef } from "../motion/velocity.js";
 import { AuthorityDef, MeDef } from "../net/components.js";
-import { ColliderDef } from "../physics/collider.js";
+import { ColliderDef, ColliderFromMeshDef } from "../physics/collider.js";
 import { WorldFrameDef } from "../physics/nonintersection.js";
 import { PositionDef, RotationDef } from "../physics/transform.js";
 import { RenderableConstructDef } from "../render/renderer-ecs.js";
@@ -29,11 +29,12 @@ const MastObj = defineObj({
   name: "mast",
   components: [
     RenderableConstructDef,
-    ColliderDef,
+    // ColliderDef,
+    ColliderFromMeshDef,
     PositionDef,
     RotationDef,
     ColorDef,
-    AuthorityDef,
+    // AuthorityDef,
   ],
   propsType: T<{ force: number }>(),
   physicsParentChildren: true,
@@ -53,30 +54,22 @@ export const HasMastObj = defineObj({
 } as const);
 export const HasMastDef = HasMastObj.props;
 
-export function createMast(
-  res: Resources<[typeof MeDef, typeof MastMesh.def]>
-) {
+export function createMast() {
   const sailWidth = 14;
   const sail = createSail(sailWidth, 8, 2);
   sail.position[0] = -sailWidth;
   sail.position[1] = 0.51;
   sail.position[2] = 38;
 
-  const mesh = res.mesh_mast;
-
   const ent = MastObj.new({
     props: { force: 0.0 },
     args: {
-      renderableConstruct: [mesh.proto],
-      collider: {
-        shape: "AABB",
-        solid: false,
-        aabb: mesh.aabb,
-      },
+      renderableConstruct: [MastMesh],
+      colliderFromMesh: false,
       position: undefined,
       rotation: undefined,
       color: ENDESGA16.darkBrown,
-      authority: res.me.pid,
+      // authority: res.me.pid,
     },
     children: {
       sail: sail,
